Parse crazy commit messages (#10)
My silly old way of parsing commit messages for version-bump directives just echoed the commit and piped it into a script. That would break, of course, if the commit message had anything that should have been escaped — like a single quote. I've changed it to just store the message in an environment variable, and have the script read the message from the environment.