Add QUATERNIONIC_DISABLE_CACHE environment variable.
It may not always be desirable to cache the Numba-generated code, for
example in a HPC environment where having lots of workers each compiling
the code may be preferable to them all trying to load it from the cache.

This adds a QUATERNIONIC_DISABLE_CACHE environment variable. If set to
1, the `cache` parameter of the `jit` and `guvectorize` decorators is
set to False. Default behaviour remains the same as before, i.e., to
use the cache.