Add broadcasting rotation method.
This uses standard NumPy broadcasting rules. This means that for an Nx4
quaternionic array and an Nx3 vector array, an Nx3 rotated vector array
will be output. This is in contrast to the existing rotate() method
which operates as an outer product and would return an NxNx3 rotated
vector array.